requirements.txt
config==0.5.1
anthropic==0.29.0
twilio==9.2.1
google-cloud-texttospeech==2.16.3
google-cloud-speech==2.26.0
pyaudio==0.2.14